+++ docs/v0.3.0/CHANGELOG.md
@@ -1,9 +1,11 @@
 # CHANGELOG
 
-# 0.3.0 <small><sup>21-09-2022</sup></small>
+# 0.3.0 <small><sup>22-09-2022</sup></small>
 
 - [Updated] enhanced default configs in file manager
 - [Updated] enhanced individual file configs in file manager
+- [Updated] the makeHandler can take one or more handlers
+- [Updated] improved on overall typings
 
 # 0.2.2 <small><sup>16-09-2022</sup></small>